+++ database/migrations/2023_02_28_103450_create_payments_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->string('paymentType')->nullable();
             $table->string('paymentTransectionID')->nullable();
             $table->string('paymentCardID')->nullable();
-            $table->double('deduction')->nullable();
-            $table->double('price_after_deduction')->nullable();
+            $table->decimal('deduction', 10, 2)->nullable();
+            $table->decimal('price_after_deduction', 10, 2)->nullable();
             $table->integer('default_option')->nullable();
             $table->unsignedBigInteger('orderID')->nullable();
             $table->foreign('orderID')->references('id')->on('orders')->onDelete('cascade');
